chore: Fix lint about hot reload
