+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react"
-import { useAuth } from "./contexts/AuthContext"
+import { useAuth } from "./hooks/useAuth"
 import { Login } from "./components/Login"
 import { SignUp } from "./components/SignUp"
