+++ frontend/src/components/SignUp.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../hooks/useAuth'
 
 export function SignUp() {
   const [username, setUsername] = useState('')
